add command registration for future commands update long command descriptions split command-specific code out of the root command
