--- pkg/cli/types.go
@@ -1,3 +1,0 @@
-package cli
-
-const TerrascanHomeDirName = ".terrascan"
